requirements.txt
langgraph==0.2.16
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
vaderSentiment==3.3.2
textblob==0.17.1
serpapi==0.1.5
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6